refactor: move background map service into visualization package
